requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.8
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1
